+++ ChangeLog.md
@@ -13,6 +13,11 @@
     (Roland Denis, [#359](https://github.com/DGtal-team/DGtalTools/pull/359/files))
   - Using SourceForge to download doxygen sources during Travis CI jobs.
     (Roland Denis [#360](https://github.com/DGtal-team/DGtalTools/pull/360))
+
+  - volAddBorder: Add an option that sets zero value to domain boundary voxels without 
+    changing the domain extent (Bertrand Kertautret 
+    [#371](https://github.com/DGtal-team/DGtalTools/pull/371))
+
   - Fix a wrong error message that appears when using the tool (wrong IO error)
     (Bertrand Kerautret
     [#368](https://github.com/DGtal-team/DGtalTools/pull/368))
@@ -21,7 +26,6 @@
   -  Fix doc link to the DGtal lib in the tool source (from new github website).
      (Bertrand Kerautret [#372](https://github.com/DGtal-team/DGtalTools/pull/372))
     
-
 - *converters*
   - itk2vol: change the type of the threshold parameter in order to be able to
     convert ITK images of type double, it also adds a new option to mask the